# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReviewPlate API - Smart NFC/QR review card platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /support/messages:
    get:
      operationId: listSupportMessages
      tags: [support]
      summary: List support messages for current user
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportMessage"

    post:
      operationId: createSupportMessage
      tags: [support]
      summary: Send a support message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessage"

  # SUBSCRIPTIONS
...
  /subscriptions/upgrade:
    post:
      operationId: upgradePlan
      tags: [subscriptions]
      summary: Upgrade subscription plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpgradePlanBody"
      responses:
        "200":
          description: Upgrade initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradePlanResponse"

  # PUBLIC ENDPOINTS (no auth)
...
components:
  schemas:
...
    SupportMessage:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        messageText:
          type: string
        sentDate:
          type: string
        isRead:
          type: boolean
        repliedByAdmin:
          type: ["string", "null"]
        replyDate:
          type: ["string", "null"]
      required:
        - id
        - senderId
        - messageText
        - sentDate
        - isRead
    CreateSupportMessageBody:
      type: object
      properties:
        messageText:
          type: string
      required:
        - messageText
...
    UpgradePlanBody:
      type: object
      properties:
        plan:
          type: string
      required:
        - plan
    UpgradePlanResponse:
      type: object
      properties:
        success:
          type: boolean
        plan:
          type: string
      required:
        - success
        - plan
...